+++ Launchpad/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.dji.simulatorDemo">
+    package="com.dji.launchpad">
 
     <!-- SDK permission requirement -->
     <uses-permission android:name="android.permission.BLUETOOTH" />
@@ -27,7 +27,7 @@
     <!-- SDK requirement permission end -->
 
     <application
-        android:name=".MApplication"
+        android:name="com.dji.launchpad.MApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -41,14 +41,14 @@
         -->
         <meta-data
             android:name="com.dji.sdk.API_KEY"
-            android:value="e2f0bfded23cd91be56d83c0" />
+            android:value="a221b174ee100cb1dc60a666" />
 
         <service android:name="dji.sdk.sdkmanager.DJIGlobalService" >
         </service>
 
         <uses-library android:name="org.apache.http.legacy" android:required="false" />
 
-        <activity android:name=".MainActivity"
+        <activity android:name="com.dji.launchpad.MainActivity"
             android:configChanges="orientation"
             android:screenOrientation="fullSensor" >
             <intent-filter>
